Fixes two compatibility issues with newer version of pyfits and numpy.  (Not needed now, but they're in for when we update.)
git-svn-id: http://chara.hi.gemini.edu/svn/DRSoftware/gemini_python/trunk@5853 e04282df-fb77-4b70-bf7a-c1223586725c

+++ astrodata_Gemini/RECIPES_Gemini/primitives/primitives_GMOS_IMAGE.py
@@ -424,7 +424,7 @@ class GMOS_IMAGEPrimitives(GMOSPrimitives):
 
             # Find the mode and standard deviation
             hist,edges = np.histogram(stat_region,
-                                      bins=np.max(sci_data)/0.1)
+                                      bins=int(np.max(sci_data)/0.1))
             mode = edges[np.argmax(hist)]
             std = np.std(stat_region)
